requirements.txt
PyQt6>=6.6
mysql-connector-python>=9.0.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
reportlab>=4.0.0
